新增个人中心与消息提醒

+++ server/api.js
@@ -88,6 +88,29 @@ module.exports = {
       })
     });
   },
+  // 获取将要到期租用车位数量
+  getRentNum(req, res, next){
+    let userId = req.query.userId;
+    let nowDate = req.query.nowDate;
+    let expireDate = req.query.expireDate;
+    pool.getConnection((err, connection) => {
+      connection.query(sqlMap.pps.queryRentNum,[userId,nowDate,expireDate],(err, result) => {
+        jsonWrite(res, result);
+        connection.release();
+      })
+    })
+  },
+  // 获取过期车位数量
+  getReleaseNum(req, res, next){
+    let userId = req.query.userId;
+    let nowDate = req.query.nowDate;
+    pool.getConnection((err, connection) => {
+      connection.query(sqlMap.pps.queryReleaseNum,[userId,nowDate],(err, result) => {
+        jsonWrite(res, result);
+        connection.release();
+      })
+    })
+  },
   // 获取所有用户
   getPpsAll(req, res, next) {
     pool.getConnection((err, connection) => {
@@ -100,8 +123,9 @@ module.exports = {
   // 获取预约车位列表
   getImagesList(req, res, next) {
     let userid = req.query.userid;
+    let nowTime = req.query.nowTime;
     pool.getConnection((err, connection) => {
-      connection.query(sqlMap.pps.queryAllPark,[userid],(err, result) => {
+      connection.query(sqlMap.pps.queryAllPark,[userid,nowTime],(err, result) => {
         jsonWrite(res, result);
         connection.release();
       })
@@ -147,20 +171,27 @@ module.exports = {
   // 搜索车位信息
   searchUserPark(req, res, next) {
     let parkcity = req.query.parkcity;
+    let startTime = req.query.startTime;
+    let endTime = req.query.endTime;
     let userid = req.query.userid;
     let section = req.query.section;
     let sql = '';
     if (section == 'MINE'){
       sql = "select * from parkInfo left join user on parkinfo.subscriber=user.id where (parkcity like '%"
-      +parkcity+"%' or parkstreet like '%"+parkcity+"%') and userid="+userid+" order by status desc";
+      +parkcity+"%' or parkstreet like '%"+parkcity+"%') and userid='"+userid+"' order by status desc";
     }else if (section == 'FIND'){
-      sql = "select * from parkInfo left join user on parkinfo.subscriber=user.id where (parkcity like '%"
-      +parkcity+"%' or parkstreet like '%"+parkcity+"%') and userid!="+userid+" and status=0";
+      if(startTime != '' && endTime != ''){
+        sql = "select * from parkInfo left join user on parkinfo.subscriber=user.id where (parkcity like '%"
+      +parkcity+"%' or parkstreet like '%"+parkcity+"%') and (starttime >='"+startTime+"' and endtime <= '"+endTime+"') and userid!='"+userid+"' and status=0";
+      }else{
+        sql = "select * from parkInfo left join user on parkinfo.subscriber=user.id where (parkcity like '%"
+      +parkcity+"%' or parkstreet like '%"+parkcity+"%') and userid!='"+userid+"' and status=0";
+      }
+      
     }else {
       sql = "select * from parkInfo left join user on parkinfo.userid=user.id where (parkcity like '%"
-      +parkcity+"%' or parkstreet like '%"+parkcity+"%') and subscriber="+userid+" order by status desc";
+      +parkcity+"%' or parkstreet like '%"+parkcity+"%') and subscriber='"+userid+"' order by status desc";
     }
-
     pool.getConnection((err, connection) => {
       connection.query(sql,(err, result) => {
         jsonWrite(res, result);
@@ -201,7 +232,9 @@ module.exports = {
             message: '服务器错误,请稍后再试!'
           });
         }else{
-          fs.unlinkSync(imageurl);
+          if (imageurl){
+            fs.unlinkSync(imageurl);
+          }
           res.json({
             status: 'SUCCESS',
             message: '删除成功!'
@@ -218,13 +251,14 @@ module.exports = {
     let parkcity = params.areaString;
     let parkstreet = params.streetString;
     let parkdetails = params.detailAddr;
+    let carseatnum = params.carseatnum;
     let starttime = params.startTime;
     let endtime = params.endTime;
     let price = params.price;
     let imageurl = params.imageUrl;
     let status = params.status;
     pool.getConnection((err, connection) => {
-      connection.query(sqlMap.pps.updatePark,[parkcity,parkstreet,parkdetails,starttime,endtime,price,imageurl,status,num],(err, result) => {
+      connection.query(sqlMap.pps.updatePark,[parkcity,parkstreet,parkdetails,carseatnum,starttime,endtime,price,imageurl,status,num],(err, result) => {
         if (err){
           res.json({
             status: 'FAIL',
@@ -253,7 +287,7 @@ module.exports = {
         }else{
           res.json({
             status: 'SUCCESS',
-            message: '成功拒绝对方预约车位!'
+            message: '成功拒绝对方租用该车位!'
           });
         }
         connection.release();
@@ -273,14 +307,51 @@ module.exports = {
         }else{
           res.json({
             status: 'SUCCESS',
-            message: '对方已成功预约该车位!'
+            message: '对方已成功租用该车位!'
           });
         }
         connection.release();
       })
     })
   },
-
+  // 获取待确认数量
+  getToBeReserve(req, res, next) {
+    let userId = req.query.userId;
+    pool.getConnection((err, connection) => {
+      connection.query(sqlMap.pps.queryReserve,[userId],(err, result) => {
+        jsonWrite(res, result);
+        connection.release();
+      })
+    })
+  },
+  // 更新用户信息
+  updateUserInfo(req, res, next) {
+    let type = req.query.updateType;
+    let value = req.query.updateValue;
+    let userId = req.query.userId;
+    let sql = '';
+    if (type == 'IMAGE'){
+      sql = "update user set headimg='"+value+"' where id='"+userId+"'";
+    }else{
+      sql = "update user set banlance="+value+" where id='"+userId+"'";
+    }
+    pool.getConnection((err, connection) => {
+      connection.query(sql,(err, result) => {
+        if (err){
+          res.json({
+            status: 'FAIL',
+            message: '服务器错误,请稍后再试!'
+          });
+        }else{
+          res.json({
+            status: 'SUCCESS',
+            message: '更新成功!'
+          });
+        }
+        connection.release();
+      })
+    })
+  },
   // 新增车位信息
   addParkInfo(req, res, next){
     let params = req.query;
@@ -289,6 +360,7 @@ module.exports = {
     let parkcity = params.areaString;
     let parkstreet = params.streetString;
     let parkdetails = params.detailAddr;
+    let carseatnum = params.carseatnum;
     let starttime = params.startTime;
     let endtime = params.endTime;
     let price = params.price;
@@ -296,7 +368,7 @@ module.exports = {
     let imageurl = params.imageUrl;
     let status = params.status;
     pool.getConnection((err, connection) => {
-      connection.query(sqlMap.pps.insertPark,[num,userid,parkcity,parkstreet,parkdetails,starttime,endtime,price,totalcost,imageurl,status],(err, result) => {
+      connection.query(sqlMap.pps.insertPark,[num,userid,parkcity,parkstreet,parkdetails,carseatnum,starttime,endtime,price,totalcost,imageurl,status],(err, result) => {
         if (err){
           fs.unlinkSync(imageurl);
           res.json({
